refactor: move postComment to its own file
Separated postComment from actions and server files for better modularity.
