+++ src/sdk/actions.ts
@@ -31,8 +31,6 @@ const inputSchema = T.Object({
   signature: T.String(),
 });
 
-const HEADER_NAME = "Ubiquity";
-
 export async function createActionsPlugin<TConfig = unknown, TEnv = unknown, TSupportedEvents extends WebhookEventName = WebhookEventName>(
   handler: (context: Context<TConfig, TEnv, TSupportedEvents>) => Promise<Record<string, unknown> | undefined>,
   options?: Options
@@ -141,46 +139,4 @@ async function returnDataToKernel(repoToken: string, stateId: string, output: ob
       output: output ? JSON.stringify(output) : null,
     },
   });
-}
-
-function createStructuredMetadata(className: string | undefined, logReturn: LogReturn) {
-  const logMessage = logReturn.logMessage;
-  const metadata = logReturn.metadata;
-
-  const jsonPretty = sanitizeMetadata(metadata);
-  const stack = logReturn.metadata?.stack;
-  const stackLine = (Array.isArray(stack) ? stack.join("\n") : stack)?.split("\n")[2] ?? "";
-  const caller = stackLine.match(/at (\S+)/)?.[1] ?? "";
-  const ubiquityMetadataHeader = `<!-- ${HEADER_NAME} - ${className} - ${caller} - ${metadata?.revision}`;
-
-  let metadataSerialized: string;
-  const metadataSerializedVisible = ["```json", jsonPretty, "```"].join("\n");
-  const metadataSerializedHidden = [ubiquityMetadataHeader, jsonPretty, "-->"].join("\n");
-
-  if (logMessage?.type === "fatal") {
-    // if the log message is fatal, then we want to show the metadata
-    metadataSerialized = [metadataSerializedVisible, metadataSerializedHidden].join("\n");
-  } else {
-    // otherwise we want to hide it
-    metadataSerialized = metadataSerializedHidden;
-  }
-
-  return metadataSerialized;
-}
-
-/**
- * Posts a comment on a GitHub issue if the issue exists in the context payload, embedding structured metadata to it.
- */
-export async function postComment(context: Context, message: LogReturn) {
-  if ("issue" in context.payload && context.payload.repository?.owner?.login) {
-    const metadata = createStructuredMetadata(message.metadata?.name, message);
-    await context.octokit.rest.issues.createComment({
-      owner: context.payload.repository.owner.login,
-      repo: context.payload.repository.name,
-      issue_number: context.payload.issue.number,
-      body: [message.logMessage.raw, metadata].join("\n"),
-    });
-  } else {
-    context.logger.info("Cannot post comment because issue is not found in the payload");
-  }
 }
