+++ src/sdk/index.ts
@@ -1,5 +1,6 @@
 export { createPlugin } from "./server";
-export { createActionsPlugin, postComment } from "./actions";
+export { createActionsPlugin } from "./actions";
+export { postComment } from "./comment";
 export type { Context } from "./context";
 export * from "./constants";
 export type { Manifest } from "../types/manifest";
